refactor: rename vault to bao module
- Rename vault/ directory to bao/ to reflect OpenBao usage
- Update package declarations from 'package vault' to 'package bao'
- All 20 Go files updated with new package name
