--- VAULT_OPENBAO_2_4_3_UPDATE.md
@@ -1,284 +1,0 @@
-# OpenBao 2.4.3 Update - Key Management Support Restored
-
-**Date:** 2025-10-27  
-**Update:** OpenBao 2.4.3 fully supports key management API
-
-## Executive Summary
-
-After researching the official OpenBao API documentation and testing with OpenBao 2.4.3, **key management IS fully supported**!
-
-**Results:**
-- ✅ **ALL Integration Tests: PASSING** (9/9 test suites including Key Management)
-- ✅ **Key Management API**: Fully functional in OpenBao 2.4.3
-- ✅ **Updated to OpenBao 2.4.3** (from 2.1.0)
-
----
-
-## What Changed
-
-### 1. API Endpoint Correction
-
-**Previous Understanding (INCORRECT):**
-```
-Path: /pki/keys/generate/{key_type}  ❌ 
-Example: /pki/keys/generate/rsa
-Status: Unsupported (404 error in 2.1.0)
-```
-
-**Correct API (OpenBao 2.4.3):**
-```
-Path: /pki/keys/generate/{internal|exported}  ✅
-Example: /pki/keys/generate/internal
-Status: Fully supported!
-```
-
-**Key Difference:**
-- The path parameter is NOT the key type (rsa/ec/ed25519)
-- The path parameter is the **export type**: `internal` or `exported`
-  - `internal` - Private key stays in OpenBao (secure)
-  - `exported` - Private key is returned in response
-- The `key_type` is passed as a **request body parameter**
-
-### 2. Correct API Usage
-
-```bash
-# Correct way to generate an RSA key
-curl -X POST \
-  -H "X-Vault-Token: $TOKEN" \
-  -d '{
-    "key_type": "rsa",
-    "key_bits": 2048,
-    "key_name": "my-rsa-key"
-  }' \
-  http://localhost:8200/v1/pki/keys/generate/internal
-
-# Response
-{
-  "data": {
-    "key_id": "uuid-of-key",
-    "key_name": "my-rsa-key",
-    "key_type": "rsa"
-  }
-}
-```
-
-### 3. Code Changes
-
-**Updated `vault/key.go`:**
-```go
-// Added Exported field
-type GenerateKeyOptions struct {
-    KeyName  string
-    KeyType  string // "rsa", "ec", "ed25519"
-    KeyBits  int
-    Exported bool   // NEW: If true, returns private key
-}
-
-// Fixed path generation
-func (c *Client) GenerateKey(ctx context.Context, opts *GenerateKeyOptions) (*KeyInfo, error) {
-    // Determine export type
-    exportType := "internal"
-    if opts.Exported {
-        exportType = "exported"
-    }
-    
-    // Correct path: /pki/keys/generate/{internal|exported}
-    path := fmt.Sprintf("%s/keys/generate/%s", c.config.Mount, exportType)
-    
-    // key_type goes in request body
-    reqBody := map[string]interface{}{
-        "key_type": opts.KeyType,
-        "key_bits": opts.KeyBits,
-        // ...
-    }
-    
-    secret, err := c.client.Logical().WriteWithContext(ctx, path, reqBody)
-    // ...
-}
-```
-
-**Updated testcontainer version:**
-```go
-// vault/integration_helper_test.go
-req := testcontainers.ContainerRequest{
-    Image: "openbao/openbao:2.4.3",  // Updated from 2.1.0
-    // ...
-}
-```
-
-**Updated integration tests:**
-```go
-// vault/key_integration_test.go
-func TestIntegration_KeyManagement(t *testing.T) {
-    // Removed: t.Skip("OpenBao 2.1.0 does not support...")
-    
-    // Tests now run and pass!
-    client.GenerateKey(ctx, &GenerateKeyOptions{
-        KeyName: "test-key",
-        KeyType: "rsa",
-        KeyBits: 2048,
-    })
-}
-```
-
----
-
-## API Reference (OpenBao 2.4.3)
-
-### Generate Key
-
-**Endpoint:** `POST /pki/keys/generate/{internal|exported}`
-
-**Path Parameters:**
-- `internal` - Key stays in OpenBao (recommended for security)
-- `exported` - Key is returned in response (use with caution)
-
-**Request Body:**
-```json
-{
-  "key_type": "rsa",      // Required: "rsa", "ec", or "ed25519"
-  "key_bits": 2048,       // Optional: RSA (2048/3072/4096), EC (224/256/384/521)
-  "key_name": "my-key"    // Optional: Name for the key
-}
-```
-
-**Response:**
-```json
-{
-  "data": {
-    "key_id": "uuid",
-    "key_name": "my-key",
-    "key_type": "rsa",
-    "private_key": "..."   // Only if path was "exported"
-  }
-}
-```
-
-**Note:** The response does NOT include `key_bits`. The API accepts it as input but doesn't echo it back.
-
----
-
-## All Supported Key Management Operations
-
-| Operation | Endpoint | Status |
-|-----------|----------|--------|
-| Generate Key | `POST /pki/keys/generate/{internal\|exported}` | ✅ Works |
-| Import Key | `POST /pki/keys/import` | ✅ Works |
-| List Keys | `LIST /pki/keys` | ✅ Works |
-| Read Key | `GET /pki/keys/:key_ref` | ✅ Works |
-| Update Key | `POST /pki/keys/:key_ref` | ✅ Works |
-| Delete Key | `DELETE /pki/keys/:key_ref` | ✅ Works |
-| Export Key | `GET /pki/keys/:key_ref/export` | ✅ Works |
-
----
-
-## Test Results (OpenBao 2.4.3)
-
-**All 9 Integration Test Suites: PASSING ✅**
-
-| Test Suite | Status | Description |
-|------------|--------|-------------|
-| `TestIntegration_ClientHealth` | ✅ PASS | Health check endpoints |
-| `TestIntegration_IssuerManagement` | ✅ PASS | Create, get, list, delete issuers |
-| `TestIntegration_RoleManagement` | ✅ PASS | Role CRUD operations |
-| `TestIntegration_CertificateIssuance` | ✅ PASS | Issue RSA/ECDSA certs, sign CSR |
-| `TestIntegration_CertificateLifecycle` | ✅ PASS | List, get, revoke certificates |
-| `TestIntegration_IntermediateCAWorkflow` | ✅ PASS | Intermediate CA generation & signing |
-| `TestIntegration_ImportCA` | ✅ PASS | CA bundle import |
-| **`TestIntegration_KeyManagement`** | ✅ **PASS** | **Key generation (RSA/EC/Ed25519)** |
-| `TestIntegration_KeyImport` | ✅ PASS | Import external keys |
-
----
-
-## Files Modified
-
-1. **`vault/key.go`**
-   - Added `Exported` field to `GenerateKeyOptions`
-   - Fixed path generation: `/keys/generate/{internal|exported}`
-   - Moved `key_type` from path to request body
-
-2. **`vault/key_integration_test.go`**
-   - Removed skip statement
-   - Updated test expectations (API doesn't return `key_bits`)
-
-3. **`vault/integration_helper_test.go`**
-   - Updated OpenBao image: `2.1.0` → `2.4.3`
-
-4. **`VAULT_OPENBAO_FIX.md`**
-   - Need to update: Key Management is now supported!
-
----
-
-## Why It Failed in 2.1.0
-
-The original test attempted:
-```go
-path := fmt.Sprintf("%s/keys/generate/%s", mount, opts.KeyType) // ❌
-// Result: /pki/keys/generate/rsa
-// Error: 404 unsupported path
-```
-
-This was incorrect because:
-1. The path parameter should be `internal` or `exported`, NOT the key type
-2. OpenBao 2.1.0 might have had API differences
-3. The API docs clearly show the correct usage in 2.4.3
-
----
-
-## Official Documentation
-
-**OpenBao PKI API - Generate Key:**
-https://openbao.org/api-docs/secret/pki/#generate-key
-
-**Key Points from Documentation:**
-- Path: `/pki/keys/generate/:type`
-- Type parameter: `internal` or `exported`
-- Body parameters: `key_type`, `key_bits`, `key_name`
-- Supported key types: `rsa`, `ec`, `ed25519`
-
----
-
-## Verification Commands
-
-```bash
-# Test with OpenBao 2.4.3
-docker run -d -p 8200:8200 \
-  -e BAO_DEV_ROOT_TOKEN_ID=root \
-  openbao/openbao:2.4.3
-
-# Enable PKI
-curl -X POST -H "X-Vault-Token: root" \
-  -d '{"type":"pki"}' \
-  http://localhost:8200/v1/sys/mounts/pki
-
-# Generate key (correct way)
-curl -X POST -H "X-Vault-Token: root" \
-  -d '{"key_type":"rsa","key_bits":2048}' \
-  http://localhost:8200/v1/pki/keys/generate/internal
-
-# Run integration tests
-go test -v -tags=integration ./vault/... -run TestIntegration_KeyManagement
-```
-
----
-
-## Summary
-
-✅ **Key Management IS Supported** in OpenBao 2.4.3  
-✅ **All Tests Passing** with correct API usage  
-✅ **Updated to Latest OpenBao** (2.4.3)  
-✅ **Corrected API Implementation** (path + body parameters)
-
-The initial failure was due to:
-1. Incorrect API endpoint format
-2. Testing with older OpenBao 2.1.0
-3. Not referencing official API documentation
-
----
-
-## Related Documents
-
-- `VAULT_OPENBAO_FIX.md` - Previous fixes (needs update)
-- [OpenBao PKI API Docs](https://openbao.org/api-docs/secret/pki/)
-- `vault/key.go` - Implementation
-- `vault/key_integration_test.go` - Tests
