+++ bao/certificate_test.go
@@ -1,4 +1,4 @@
-package vault
+package bao
 
 import (
 	"context"
