+++ bao/integration.go
@@ -1,4 +1,4 @@
-package vault
+package bao
 
 import (
 	"crypto/x509"
