+++ bao/types.go
@@ -1,7 +1,7 @@
 // Package vault provides integration with Vault/OpenBao PKI secrets engine.
 // It enables seamless interaction between GoPKI's type-safe cryptographic operations
 // and centralized PKI management via Vault/OpenBao.
-package vault
+package bao
 
 import (
 	"crypto/x509"
